requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pinecone>=7.0.0
datetime>=5.4
pydantic>=2.0.0
openai>=1.0.0
